+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-
+- Make `Start` and `Awake` in `GameCore` virtual.
+- 
 ---
 
 ## [2.0.0] - 2025-05-15