+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+- Add cursor state method to `MainSceneFlowController`.
+
 ---
 
 [2.1.0] - 2025-05-23