+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 - Add cursor state method to `MainSceneFlowController`.
+- Make OnEnable and OnDisable protected in `MainSceneFlowController`.
 
 ---