+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add cursor state method to `MainSceneFlowController`.
 - Make OnEnable and OnDisable protected in `MainSceneFlowController`.
 - Add quit game method to `IGameFlowController`.
+- Change default audio settings values to a more sensible 0.5.
 
 ---