+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Make OnEnable and OnDisable protected in `MainSceneFlowController`.
 - Add quit game method to `IGameFlowController`.
 - Change default audio settings values to a more sensible 0.5.
+- Clean up context menu items.
 
 ---